pyproject.toml
[project]
name = "magic_top_hat"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Luca Mattiazzi", email = "l.d.mattiazzi@gmail.com" }]
dependencies = ["openai>=1.35.13", "python-dotenv>=1.0.1"]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/magic_top_hat"]
